♻️ Move ConnOptions and DefaultConnOptions to opts
